# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: Task Management API
  description: |-
    タスク管理のAPI仕様書です。
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local server
tags:
  - name: login
    description: ユーザーログインに関する操作
  - name: register
    description: ユーザー情報の登録に関する操作
  - name: task
    description: タスク管理に関する操作
  - name: user
    description: ユーザー情報に関する操作
paths:
  /login:
    get:
      tags:
        - login
      summary: ログインフォームを表示
      operationId: showLoginForm
      responses:
        '200':
          description: ログインページを表示します
          content:
            text/html:
              schema:
                type: string

  /logout:
    get:
      tags:
        - login
      summary: ログアウトページを表示
      operationId: showLogout
      responses:
        '200':
          description: ログアウトページを表示します
          content:
            text/html:
              schema:
                type: string

  /register:
    get:
      tags:
        - register
      summary: 新規ユーザーの登録フォームを表示
      operationId: RegistrationForm
      responses:
        '200':
          description: 新規ユーザーの登録ページを表示します
          content:
            text/html:
              schema:
                type: string
    
    post:
      tags:
        - register
      summary: 新規ユーザーを登録
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: ユーザー名
                password:
                  type: string
                  description: パスワード
                email:
                  type: string
                  description: メールアドレス
      responses:
        '200':
          description: 登録成功またはエラーメッセージを表示します
          content:
            text/html:
              schema:
                type: string

  components:
    schemas:
      RegisterUser:
        type: object
        properties:
          username:
            type: string
          password:
            type: string
          email:
            type: string

  /task:
    get:
      tags:
        - task
      summary: 全てのタスクと完了済みタスクを取得
      operationId: listHome
      responses:
        '200':
          description: 全てのタスクと完了済みタスクのリストを返します
          content:
            text/html:
              schema:
                type: string
    
    post:
      tags:
        - task
      summary: 新しいタスクを追加
      operationId: addItem
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                client:
                  type: string
                  description: クライアント名
                contractor:
                  type: string
                  description: 請負業者名
                task:
                  type: string
                  description: タスクの説明
                deadline:
                  type: string
                  format: date
                  description: タスクの締め切り
          responses:
            '302':
          description: タスクリストページへリダイレクト
  
  /task/{taskId}:
    delete:
      tags:
        - task
      summary: 現行のタスクを削除
      operationId: deleteTask
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
          description: 削除するタスクのID
      responses:
        '302':
          description: タスクリストページへリダイレクト
    
    /task/update:
      patch:
        tags:
          - task
      summary: 現行のタスクを更新
      operationId: updateItem
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                taskId:
                client:
                  type: string
                  description: クライアント名
                contractor:
                  type: string
                  description: 請負業者名
                task:
                  type: string
                  description: タスクの説明
                deadline:
                  type: string
                  format: date
                  description: タスクの締め切り
                done:
                  type: boolean
                  description: タスクが完了しているかどうか
      responses:
        '302':
          description: タスクリストページへリダイレクト

  /task/{taskId}/complete:
    post:
      tags:
        - task
      summary: タスクを完了としてマーク
      operationId: completeTask
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
          description: 完了としてマークするタスクのID
      responses:
        '302':
          description: タスクリストページへリダイレクト

  /task/{taskId}/completedTask:
    post:
      tags:
        - task
      summary: 作業完了済みタスクを履歴リストに移動
      operationId: completedTask
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
          description: 履歴リストに移動するタスクのID
      responses:
        '302':
          description: タスクリストページへリダイレクト

  /task/{taskId}/notCompletedTask:
    post:
      tags:
        - task
      summary: 完了済みタスクをアクティブリストに戻す
      operationId: notCompletedTask
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
          description: アクティブリストに戻すタスクのID
      responses:
        '302':
          description: タスクリストページへリダイレクト
components:
  schemas:
    TaskItem:
      type: object
      properties:
        taskId:
          type: string
        client:
          type: string
        contractor:
          type: string
        task:
          type: string
        deadline:
          type: string
          format: date
        done:
          type: boolean
    CompTaskItem:
      type: object
      properties:
        taskId:
          type: string
        client:
          type: string
        contractor:
          type: string
        task:
          type: string
        deadline:
          type: string
          format: date
        completedDate:
          type: string
          format: date-time
    RecordTaskItem:
      type: object
      properties:
        taskId:
          type: string
        client:
          type: string
        contractor:
          type: string
        task:
          type: string
        deadline:
          type: string
          format: date
        completedDate:
          type: string
          format: date-time